adding video audio call mode

+++ src/components/message/RightSide.js
@@ -33,7 +33,7 @@ const RightSide = () => {
   const [page, setPage] = useState(0);
   const [isLoadMore, setIsLoadMore] = useState(false);
 
-  const history = useHistory()
+  const history = useHistory();
 
   useEffect(() => {
     const newData = message.data.find((item) => item._id === id);
@@ -155,8 +155,32 @@ const RightSide = () => {
   }, [auth, id, page, dispatch, isLoadMore]);
 
   const handleDeleteConversation = () => {
-    dispatch(deleteConversation({auth, id}))
-    return history.push("/message")
+    if (window.confirm("Do you want to delete?")) {
+      dispatch(deleteConversation({ auth, id }));
+      return history.push("/message");
+    }
+  };
+
+  // Call
+  const caller = ({ video }) => {
+    const { _id, avatar, username, fullname } = user;
+    const msg = {
+      sender: auth.user._id,
+      recipient: _id,
+      avatar,
+      username,
+      fullname,
+      video,
+    };
+    dispatch({ type: GLOBALTYPES.CALL, payload: msg });
+  };
+
+  const handleAudioCall = () => {
+    caller({ video: false });
+  };
+
+  const handleVideoCall = () => {
+    caller({ video: true });
   };
 
   return (
@@ -164,11 +188,23 @@ const RightSide = () => {
       <div className="message-header">
         {user.length !== 0 && (
           <UserCard user={user}>
-            <i
-              className="fas fa-trash text-danger"
-              style={{ cursor: "pointer" }}
-              onClick={handleDeleteConversation}
-            />
+            <div>
+              <i
+                className="fas fa-phone-alt"
+                style={{ cursor: "pointer" }}
+                onClick={handleAudioCall}
+              />
+              <i
+                className="fas fa-video mx-3"
+                style={{ cursor: "pointer" }}
+                onClick={handleVideoCall}
+              />
+              <i
+                className="fas fa-trash text-danger"
+                style={{ cursor: "pointer" }}
+                onClick={handleDeleteConversation}
+              />
+            </div>
           </UserCard>
         )}
       </div>
